pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "careermate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12, <=3.13"
dependencies = [
    "django==4.2.16",
    "djangorestframework==3.16.1",
    "huggingface-hub==0.35.3",
    "langchain==0.3.27",
    "langchain-core==0.3.79",
    "langchain-openai==0.3.35",
    "langchain-text-splitters==0.3.11",
    "langchain-weaviate==0.0.5",
    "langchain-community",
    "langcodes==3.5.0",
    "langsmith==0.4.34",
    "language-data==1.3.0",
    "numpy==2.3.3",
    "openai==2.3.0",
    "pandas==2.3.3",
    "pillow==11.3.0",
    "PyJWT",
    "scikit-learn==1.7.2",
    "sentence-transformers==5.1.1",
    "tokenizers==0.22.1",
    #    "torch",
    #    "torchvision",
    #    "torchaudio",
    "torchmetrics==1.8.2",
    "tqdm==4.67.1",
    "weaviate-client==4.17.0",
    "transformers==4.57.0",
    "jupyter",
    "notebook",
    # swagger/openapi for DRF
    "drf-spectacular>=0.28.0",
    "drf-spectacular-sidecar>=2025.10.1",
    # Gemini
    "langchain-google-genai==2.0.10",
    "google-generativeai==0.8.5",
    "google-genai",
    "langchain-ollama>=0.3.10",
    "pypdf2>=3.0.1",
    "pdfplumber>=0.11.7",
    "pymupdf>=1.26.5",
    "docx>=0.2.4",
    "pdf2image>=1.17.0",
    "pytesseract>=0.3.13",
    "poppler-utils>=0.1.0",
    "redis>=6.4.0",
    "celery==5.5.3",
    "traits>=7.0.2",
    "docx2txt>=0.9",
    "psycopg2-binary>=2.9.11",
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "asgiref==3.10.0",
    "authlib==1.6.5",
    "black==25.9.0",
    "blis==1.3.0",
    "cachetools==6.2.1",
    "catalogue==2.0.10",
    "certifi==2025.10.5",
    "cffi==2.0.0",
    "charset-normalizer==3.4.3",
    "click==8.3.0",
    "cloudpathlib==0.23.0",
    "colorama==0.4.6",
    "confection==0.1.5",
    "cryptography==46.0.2",
    "cymem==2.0.11",
    "deprecation==2.1.0",
    "distro==1.9.0",
    "drf-yasg==1.21.11",
    "filelock==3.20.0",
    "filetype==1.2.0",
    "fsspec==2025.9.0",
    "google-ai-generativelanguage==0.6.15",
    "google-api-core==2.26.0",
    "google-api-python-client==2.184.0",
    "google-auth==2.41.1",
    "google-auth-httplib2==0.2.0",
    "googleapis-common-protos==1.70.0",
    "greenlet==3.2.4",
    "grpcio==1.75.1",
    "grpcio-status==1.71.2",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httplib2==0.31.0",
    "httpx==0.28.1",
    "idna==3.11",
    "inflection==0.5.1",
    "iniconfig==2.1.0",
    "isort==7.0.0",
    "jinja2==3.1.6",
    "jiter==0.11.0",
    "joblib==1.5.2",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "lightning-utilities==0.15.2",
    "marisa-trie==1.3.1",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "mdurl==0.1.2",
    "mpmath==1.3.0",
    "murmurhash==1.0.13",
    "mypy-extensions==1.1.0",
    "networkx==3.5",
    "nltk==3.9.2",
    "orjson==3.11.3",
    "packaging==25.0",
    "pathspec==0.12.1",
    "pip==24.3.1",
    "platformdirs==4.5.0",
    "pluggy==1.6.0",
    "preshed==3.0.10",
    "proto-plus==1.26.1",
    "protobuf==5.29.5",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycparser==2.23",
    "pydantic==2.12.0",
    "pydantic-core==2.41.1",
    "pygments==2.19.2",
    "pyparsing==3.2.5",
    "pytest==8.4.2",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "pytokens==0.1.10",
    "pytz==2025.2",
    "pyyaml==6.0.3",
    "regex==2025.9.18",
    "requests==2.32.5",
    "requests-toolbelt==1.0.0",
    "rich==14.2.0",
    "rsa==4.9.1",
    "safetensors==0.6.2",
    "scipy==1.16.2",
    "setuptools==80.9.0",
    "shellingham==1.5.4",
    "simsimd==6.5.3",
    "six==1.17.0",
    "smart-open==7.3.1",
    "sniffio==1.3.1",
    "spacy==3.8.7",
    "spacy-legacy==3.0.12",
    "spacy-loggers==1.0.5",
    "sqlalchemy==2.0.44",
    "sqlparse==0.5.3",
    "srsly==2.5.1",
    "sympy==1.14.0",
    "tenacity==9.1.2",
    "textblob==0.19.0",
    "thinc==8.3.6",
    "threadpoolctl==3.6.0",
    "tiktoken==0.12.0",
    "torch==2.8.0",
    "typer==0.19.2",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.2",
    "uritemplate==4.2.0",
    "urllib3==2.5.0",
    "validators==0.35.0",
    "wasabi==1.1.3",
    "weasel==0.4.1",
    "wrapt==1.17.3",
    "zstandard==0.25.0",
    "faiss-cpu>=1.12.0",
    "faker>=37.12.0",
    "unsloth>=2025.7.2",
    "datasets>=4.3.0",
    "trl>=0.23.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = [
    "agent_core",
    "agent_core.*",
    "careermate",
    "Careermate.*",
    "home",
    "home.*",
    "apps",
    "apps.*",
]
exclude = ["**/tests*", "**/__pycache__*"]

[tool.uv.workspace]
members = [
    "be-python",
]
